backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
authlib==1.3.1
